feat: create component for date box
